clientID no back

+++ Backend/ReservaCruzeiros/src/main/java/ReservaCruzeiros/Pagamento/PagamentoController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class PagamentoController {
 
     @PostMapping("/pagamento/gerarLink")
-    public ResponseEntity<String> chamarOutroEndpoint(@RequestParam("idReserva") int idReserva) {
+    public ResponseEntity<String> chamarOutroEndpoint(@RequestParam("idReserva") long idReserva) {
         return ResponseEntity.ok("Link para pagamento: https://ReservaCruzeiros.com/reserva/pagamento/" + idReserva);
     }
 }
